# lib/api-zod/src/generated/types/createCharacterRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Attention Dragons API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateCharacterRequest {
  partyId?: number | null;
  name: string;
  playerName: string;
  characterClass: string;
  race: string;
  level: number;
  summary?: string | null;
  avatarUrl?: string | null;
}