# lib/api-zod/src/generated/types/createItemRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Attention Dragons API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateItemRequestCategory } from "./createItemRequestCategory";
import type { CreateItemRequestLocation } from "./createItemRequestLocation";
import type { CreateItemRequestRarity } from "./createItemRequestRarity";
import type { CreateItemRequestRechargeOn } from "./createItemRequestRechargeOn";

export interface CreateItemRequest {
  name: string;
  category: CreateItemRequestCategory;
  description: string;
  notes?: string | null;
  imageUrl?: string | null;
  quantity?: number;
  location?: CreateItemRequestLocation;
  isEquipped?: boolean;
  maxCharges?: number | null;
  currentCharges?: number | null;
  rechargeOn?: CreateItemRequestRechargeOn;
  rarity?: CreateItemRequestRarity;
  isConsumable: boolean;
}