# lib/api-zod/src/generated/types/inventoryItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Attention Dragons API specification
 * OpenAPI spec version: 0.1.0
 */
import type { InventoryItemCategory } from "./inventoryItemCategory";
import type { InventoryItemLocation } from "./inventoryItemLocation";
import type { InventoryItemRarity } from "./inventoryItemRarity";
import type { InventoryItemRechargeOn } from "./inventoryItemRechargeOn";

export interface InventoryItem {
  id: number;
  characterId: number;
  name: string;
  category: InventoryItemCategory;
  description: string;
  notes?: string | null;
  imageUrl?: string | null;
  quantity: number;
  location: InventoryItemLocation;
  isEquipped: boolean;
  maxCharges?: number | null;
  currentCharges?: number | null;
  rechargeOn?: InventoryItemRechargeOn;
  rarity?: InventoryItemRarity;
  isConsumable: boolean;
  isConsumed: boolean;
  isTrashed: boolean;
  createdAt: Date;
  updatedAt: Date;
}