# lib/api-zod/src/generated/types/updateItemRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Attention Dragons API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateItemRequestCategory } from "./updateItemRequestCategory";
import type { UpdateItemRequestLocation } from "./updateItemRequestLocation";
import type { UpdateItemRequestRarity } from "./updateItemRequestRarity";
import type { UpdateItemRequestRechargeOn } from "./updateItemRequestRechargeOn";

export interface UpdateItemRequest {
  name?: string;
  category?: UpdateItemRequestCategory;
  description?: string;
  notes?: string | null;
  imageUrl?: string | null;
  quantity?: number;
  location?: UpdateItemRequestLocation;
  isEquipped?: boolean;
  maxCharges?: number | null;
  currentCharges?: number | null;
  rechargeOn?: UpdateItemRequestRechargeOn;
  rarity?: UpdateItemRequestRarity;
  isConsumable?: boolean;
}